address review: split loading primitives into one file per component
