+++ clients/web/src/domains/settings/components/skeleton-lines.test.tsx
@@ -1,35 +1,9 @@
 import { afterEach, describe, expect, test } from "bun:test";
 import { cleanup, render } from "@testing-library/react";
 
-import {
-  ContentReveal,
-  SkeletonCardBlock,
-  SkeletonLines,
-} from "./billing-loading";
+import { SkeletonLines } from "./skeleton-lines";
 
 afterEach(cleanup);
-
-describe("ContentReveal", () => {
-  test("renders its children", () => {
-    const { getByText } = render(
-      <ContentReveal>
-        <p>Resolved content</p>
-      </ContentReveal>,
-    );
-    expect(getByText("Resolved content")).toBeDefined();
-  });
-
-  test("passes the className through to the wrapper", () => {
-    const { container } = render(
-      <ContentReveal className="flex flex-col gap-4">
-        <span>Body</span>
-      </ContentReveal>,
-    );
-    expect(container.firstElementChild?.className).toContain(
-      "flex flex-col gap-4",
-    );
-  });
-});
 
 describe("SkeletonLines", () => {
   test("renders one shimmer row per requested line", () => {
@@ -60,26 +34,14 @@ describe("SkeletonLines", () => {
 
   test("applies the line and container classNames", () => {
     const { container, getByRole } = render(
-      <SkeletonLines lines={1} lineClassName="h-10 rounded-lg" className="py-2" />,
+      <SkeletonLines
+        lines={1}
+        lineClassName="h-10 rounded-lg"
+        className="py-2"
+      />,
     );
     expect(getByRole("status").className).toContain("py-2");
     const row = container.querySelector('[data-slot="skeleton"]');
     expect(row?.className).toContain("h-10 rounded-lg");
-  });
-});
-
-describe("SkeletonCardBlock", () => {
-  test("renders its shimmer rows inside a card", () => {
-    const { container } = render(<SkeletonCardBlock />);
-    const card = container.querySelector('[data-slot="card"]');
-    expect(card).not.toBeNull();
-    expect(card?.querySelectorAll('[data-slot="skeleton"]').length).toBe(3);
-  });
-
-  test("labels the block when a label is given", () => {
-    const { getByRole } = render(<SkeletonCardBlock label="Loading billing" />);
-    expect(getByRole("status").getAttribute("aria-label")).toBe(
-      "Loading billing",
-    );
   });
 });
